[Add] User Database Acess + User Management Page + Folders in "Pages" and "Database" for Structure
