+++ SPFAdminSystem/SPFAdminSystem/Database/UserFiles/IUserData.cs
@@ -1,0 +1,12 @@
+﻿using DataAccessLibrary.Models;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace SPFAdminSystem.Database.UserFiles
+{
+    public interface IUserData
+    {
+        Task<List<User>> GetUsers();
+        Task InsertUser(User user);
+    }
+}
