+++ SPFAdminSystem/SPFAdminSystem/Pages/BlazorTemplatePages/Index.razor
