+++ SPFAdminSystem/SPFAdminSystem/Shared/NavMenu.razor
@@ -11,14 +11,14 @@
     <nav class="flex-column h-100">
         <AuthorizeView Roles="Admin, Employee">
             <div class="nav-item px-3">
-                <NavLink class="nav-link d-flex flex-row" href="product" Match="NavLinkMatch.All">
+                <NavLink class="nav-link d-flex flex-row" href="products" Match="NavLinkMatch.All">
                     <span class="oi oi-home" aria-hidden="true"></span> Product Administration
                 </NavLink>
             </div>
         </AuthorizeView>
         <AuthorizeView Roles="Admin">
             <div class="nav-item px-3">
-                <NavLink class="nav-link" href="admin">
+                <NavLink class="nav-link d-flex flex-row" href="users" Match="NavLinkMatch.All">
                     <span class="oi oi-people" aria-hidden="true"></span> User Management
                 </NavLink>
             </div>
